+++ TritonPlayerSDK/Classes/Ads/TDInterstitialAd.m
@@ -18,6 +18,7 @@
 #import "TDAnalyticsTracker.h"
 #import "AVKit/AVkit.h"
 
+
 #define kCloseButtonWidth 30
 #define kCloseButtonHeight 30
 #define kCloseButtonXPosition 10