+++ xcode_build.sh
@@ -11,7 +11,7 @@ fi
 
 echo "info: Building sdk version $SDK_VERSION"
 #xcodebuild -scheme TritonPlayerSDKStatic-Universal OTHER_CFLAGS="-fembed-bitcode" SYMROOT="Framework" -configuration Release -UseModernBuildSystem=NO
-xcodebuild -scheme TritonPlayerSDKStatic-Universal SYMROOT="Framework" -configuration Release -UseModernBuildSystem=NO SUPPORTS_MACCATALYST=NO
+xcodebuild -scheme TritonPlayerSDKStatic-Universal SYMROOT="Framework" -configuration Release -UseModernBuildSystem=YES
 
 echo "info: Generating doc"
 xcodebuild -scheme Documentation SYMROOT="Framework" -configuration Release -UseModernBuildSystem=YES SUPPORTS_MACCATALYST=NO